Implement reviewer comments
